Fix: move around router and add folder structure
