+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/linesmerrill/go-chi-template
 
-go 1.22.5
+go 1.23
 
 require github.com/go-chi/chi/v5 v5.2.0
